+++ app/src/main/java/com/example/eventplanner/MainActivity.java
@@ -1,65 +1,36 @@
 package com.example.eventplanner;
 
-import androidx.annotation.NonNull;
+
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
-import android.widget.ImageButton;
+
 
 public class MainActivity extends AppCompatActivity {
 
-    ImageButton edit1;
-    ImageButton edit2;
+    Button btn;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        getSupportActionBar().setTitle(R.string.e_home_name);
-        edit1 = (ImageButton) findViewById(R.id.edit);
-        edit2 = (ImageButton) findViewById(R.id.imageButton23);
+        getSupportActionBar().hide();
+        btn = findViewById(R.id.button);
 
     }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        MenuInflater inflater = getMenuInflater();
-        inflater.inflate(R.menu.menu_home,menu);
-        return true;
-
-    }
-
-
 
     @Override
     protected void onResume() {
         super.onResume();
-
-
-        edit2.setOnClickListener(new View.OnClickListener() {
+        btn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-
-                Intent intent = new Intent(MainActivity.this,GuestHome.class);
+                Intent intent = new Intent(MainActivity.this,Home.class);
                 startActivity(intent);
-
-            }
-        });
-
-        edit1.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-
-                Intent intent = new Intent(MainActivity.this,EventHome.class);
-                startActivity(intent);
-
             }
         });
     }
